+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Allow cancel fee/fine as error only if any 'actions' has not been applied. Refs UIU-1894.
 * Allow search by barcode. Refs UIU-1708.
 * Show the number of open requests in `LoanActionDialog`. Refs UIU-1890.
+* Add column 'Expiration date offset (days)' to patron group table. Refs UIU-1908.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)